Added example of DataframeFunctionTransformer
Please note that this is not the recommended way to do one hot encoding.
Please use something like scikit-learn's OneHotEncoder which has a whole
host of added benefits.  I have only implemented this to show how you
can integrate a pandas function into a scikit learn pipeline.  The
requirement of the function is that it must take in a dataframe and
return a dataframe.